refactor(handler): split auth, oauth, tokens files under 300 line limit
